requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
gunicorn==21.2.0
Jinja2==3.1.4
itsdangerous==2.1.2
Werkzeug==3.0.2